requirements.txt
git+https://github.com/snarfed/oauth-dropins.git#egg=oauth_dropins
git+https://github.com/snarfed/granary.git#egg=granary
git+https://github.com/snarfed/lexrpc.git#egg=lexrpc
git+https://github.com/dvska/gdata-python3.git#egg=gdata

attrs==22.2.0
beautifulsoup4==4.11.1
brevity==0.2.17
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
domain2idna==1.12.0
ecdsa==0.18.0
extras==1.0.0
feedgen==0.9.0
feedparser==6.0.10
fixtures==4.0.1
Flask==2.2.2
Flask-Caching==2.0.1
flask-gae-static==1.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-appengine-logging==1.2.0
google-cloud-audit-log==0.2.4
google-cloud-core==2.3.2
google-cloud-datastore==2.11.0
google-cloud-logging==3.3.1
google-cloud-ndb==2.1.0
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
gunicorn==20.1.0
html2text==2020.1.16
html5lib==1.1
humanfriendly==10.0
humanize==4.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.1.1
mf2py==1.1.2
mf2util==0.5.1
mox3==1.1.0
oauthlib==3.2.2
packaging==22.0
pbr==5.11.0
praw==7.6.1
prawcore==2.3.0
proto-plus==1.22.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.16.0
pymemcache==4.0.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-tumblpy==1.1.4
pytz==2022.7
PyYAML==6.0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
testtools==2.5.0
tlslite-ng==0.7.6
tweepy==4.12.1
ujson==5.6.0
update-checker==0.18.0
urllib3==1.26.13
webapp2==3.0.0b1
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.4.2
Werkzeug==2.2.2
wrapt==1.14.1
